java: clean api
