+++ java/src/main/java/com/dataseries/Union.java
@@ -2,11 +2,184 @@ package com.dataseries;
 
 import java.util.Iterator;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
 public final class Union<P extends Comparable<P>, R, L, T> implements Iterator<DataPoint<P, T>> {
+
+    sealed interface Value<T extends Comparable<T>> extends Comparable<Value<T>> permits Value.Fixed, Value.Infinite {
+
+        public static final record Fixed<T extends Comparable<T>>(T value) implements Value<T> {
+        }
+
+        public static final record Infinite<T extends Comparable<T>>() implements Value<T> {
+        }
+
+        public static <T extends Comparable<T>> Fixed<T> fixed(final T value) {
+            return new Fixed<>(value);
+        }
+
+        public static <T extends Comparable<T>> Infinite<T> infinite() {
+            return new Infinite<>();
+        }
+
+        @Override
+        default int compareTo(final Value<T> o) {
+            switch (this) {
+                case Infinite<T> i -> {
+                    switch (o) {
+                        case Infinite<T> i2 -> {
+                            return 0;
+                        }
+                        case Fixed<T> v -> {
+                            return 1;
+                        }
+                    }
+                }
+                case Fixed<T> v -> {
+                    switch (o) {
+                        case Infinite<T> i -> {
+                            return -1;
+                        }
+                        case Fixed<T> v2 -> {
+                            return v.value.compareTo(v2.value);
+                        }
+                    }
+                }
+            }
+        }
+
+        default boolean isGreaterThan(final Value<T> o) {
+            return this.compareTo(o) > 0;
+        }
+
+        default boolean isLessThan(final Value<T> o) {
+            return this.compareTo(o) < 0;
+        }
+    }
+
+    static sealed interface Cursor<T> permits Cursor.Single, Cursor.Pair {
+        static final record Single<T>(T v) implements Cursor<T> {
+        }
+
+        static final record Pair<T>(T first, T second) implements Cursor<T> {
+        }
+
+        static <T> Single<T> single(final T v) {
+            return new Single<>(v);
+        }
+
+        static <T> Pair<T> pair(final T fst, final T snd) {
+            return new Pair<T>(fst, snd);
+        }
+
+        default <R> Cursor<R> map(final Function<T, R> f) {
+            switch (this) {
+                case final Single<T> s -> {
+                    return Cursor.single(f.apply(s.v));
+                }
+                case final Pair<T> p -> {
+                    return Cursor.pair(f.apply(p.first), f.apply(p.second));
+                }
+            }
+        }
+
+        static <T extends Comparable<T>> Boolean canOverlap(final Cursor<T> left, final Cursor<T> right) {
+            final var fst = left.fst().compareTo(right.fst()) > 0 ? left.fst() : right.fst();
+            final var snd = snd(left).compareTo(snd(right)) < 0 ? snd(left) : snd(right);
+            return snd.isGreaterThan(Value.fixed(fst));
+        }
+
+        default T fst() {
+            switch (this) {
+                case final Single<T> s -> {
+                    return s.v;
+                }
+                case final Pair<T> p -> {
+                    return p.first;
+                }
+            }
+        }
+
+        static <T extends Comparable<T>> Value<T> snd(final Cursor<T> x) {
+            switch (x) {
+                case final Single<T> s -> {
+                    return Value.infinite();
+                }
+                case final Pair<T> p -> {
+                    return Value.fixed(p.second);
+                }
+            }
+        }
+    }
+
+    static class CursorIterator<T> implements Iterator<Cursor<T>> {
+
+        private final Iterator<T> iterator;
+
+        private Optional<Cursor<T>> state = Optional.empty();
+        private Boolean isPulled = false;
+        private Boolean hasNext = true;
+
+        public CursorIterator(final Iterator<T> iterator) {
+            this.iterator = iterator;
+        }
+
+        private final Optional<Cursor<T>> getState() {
+            if (this.state.isPresent()) {
+                switch (this.state.get()) {
+                    case final Cursor.Single<T> single -> Optional.empty();
+                    case final Cursor.Pair<T> pair -> {
+                        if (!this.iterator.hasNext())
+                            return Optional.of(Cursor.single(pair.second()));
+                        return Optional.of(Cursor.pair(pair.second(), this.iterator.next()));
+                    }
+                }
+            }
+
+            if (!this.iterator.hasNext())
+                return Optional.empty();
+            final var current = this.iterator.next();
+
+            if (!this.iterator.hasNext())
+                return Optional.of(Cursor.single(current));
+
+            return Optional.of(Cursor.pair(current, this.iterator.next()));
+        }
+
+        private final void pull() {
+            if (this.isPulled)
+                return;
+
+            if (!this.hasNext) {
+                this.isPulled = true;
+                return;
+            }
+
+            this.isPulled = true;
+            this.state = getState();
+            this.hasNext = this.state.isPresent();
+        }
+
+        @Override
+        public final boolean hasNext() {
+            pull();
+            return this.hasNext;
+        }
+
+        @Override
+        public final Cursor<T> next() {
+            pull();
+            if (this.state.isEmpty())
+                throw new NoSuchElementException();
+
+            this.isPulled = false;
+            return this.state.get();
+        }
+    }
+
     public static sealed interface UnionResult<L, R>
             permits UnionResult.LeftOnly, UnionResult.RightOnly, UnionResult.Both {
         public static record Both<L, R>(L left, R right) implements UnionResult<L, R> {
@@ -50,23 +223,23 @@ public final class Union<P extends Comparable<P>, R, L, T> implements Iterator<D
         static record Overlapped<L, R>(Cursor<L> left, Cursor<R> right) implements UnionState<L, R> {
         }
 
-        static <L, R> Overlapped<L, R> overlapped(final Cursor<L> left, final Cursor<R> right) {
+        private static <L, R> Overlapped<L, R> overlapped(final Cursor<L> left, final Cursor<R> right) {
             return new Overlapped<>(left, right);
         }
 
-        static <L, R> LeftOnly<L, R> leftOnly(final Cursor<L> left) {
+        private static <L, R> LeftOnly<L, R> leftOnly(final Cursor<L> left) {
             return new LeftOnly<>(left);
         }
 
-        static <L, R> RightOnly<L, R> rightOnly(final Cursor<R> right) {
+        private static <L, R> RightOnly<L, R> rightOnly(final Cursor<R> right) {
             return new RightOnly<>(right);
         }
 
-        static <L, R> Disjointed<L, R> disjointed(final Cursor<L> left, final Cursor<R> right) {
+        private static <L, R> Disjointed<L, R> disjointed(final Cursor<L> left, final Cursor<R> right) {
             return new Disjointed<>(left, right);
         }
 
-        static <L, R> None<L, R> none() {
+        private static <L, R> None<L, R> none() {
             return new None<>();
         }
