+++ java/src/test/java/com/dataseries/ValueTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import com.dataseries.Union.Value;
+
 public class ValueTest {
 
     @Test
